Fixed way of downloading script

+++ install-and-configure.sh
@@ -11,15 +11,15 @@ fi
 if [[ ! $(which python3) ]]; then
     echo "Please install python3"
 fi
-# set paths for script
-script_root_dir="$HOME/SourceVersionControls/github.com/virtualdemon"
-script_path_dir="$HOME/SourceVersionControls/github.com/virtualdemon/passwordmanager"
+# check script existence
+if [[ -e "~/.passwordmanager" ]]; then
+    echo "local repository already exist!: ~/.passwordmanager"
+    exit
+fi
 # download script repository
 echo "Downloading repository"
-mkdir -p $script_root_dir
-cd $script_root_dir
-git clone https://github.com/virtualdemon/passwordmanager
-cd $script_path_dir
+git clone https://github.com/virtualdemon/passwordmanager -o ~/.passwordmanager
+cd ~/.passwordmanager
 # install virtualenv with pip3
 pip3 install virtualenv --user
 # create .vev directory
@@ -28,24 +28,24 @@ virtualenv .venv -p python3
 if [[ $(echo $SHELL | grep bash) ]] ;then
     if [[ ! $(grep activenv ~/.bashrc) ]]; then
         echo "alias activenv='source .venv/bin/activate'" >> ~/.bashrc
-        echo "alias passwordmanager='cd ~/SourceVersionControls/github.com/virtualdemon/passwordmanager && activenv && ./passwordmanager.py'" >> ~/.bashrc
-        echo "alias update-passwordmanager='cd ~/SourceVersionControls/github.com/virtualdemon/passwordmanager && git pull -f && cd'" >> ~/.bashrc
+        echo "alias passwordmanager='cd ~/.passwordmanager && activenv && ./passwordmanager.py'" >> ~/.bashrc
+        echo "alias update-passwordmanager='cd ~/.passwordmanager && git pull -f && cd'" >> ~/.bashrc
     else
         echo "Script is configured for you!"
     fi
 elif [[ $(echo $SHELL | grep zsh) ]] ; then
     if [[ ! $(grep activenv ~/.zshrc) ]]; then
         echo "alias activenv='source .venv/bin/activate'" >> ~/.zshrc
-        echo "alias passwordmanager='cd ~/SourceVersionControls/github.com/virtualdemon/passwordmanager && activenv && ./passwordmanager.py'" >> ~/.zshrc
-        echo "alias update-passwordmanager='cd ~/SourceVersionControls/github.com/virtualdemon/passwordmanager && git pull -f && cd'" >> ~/.zshrc
+        echo "alias passwordmanager='cd ~/.passwordmanager && activenv && ./passwordmanager.py'" >> ~/.zshrc
+        echo "alias update-passwordmanager='cd ~/.passwordmanager && git pull -f && cd'" >> ~/.zshrc
     else
         echo "Script is configured for you!"
     fi
 else
     echo "****Please add these lines to your .${SHELL}rc file and continue with manual installation****"
     echo "alias activenv='source .venv/bin/activate'"
-    echo "alias passwordmanager='cd ~/SourceVersionControls/github.com/virtualdemon/passwordmanager && activenv && ./passwordmanager.py'"
-    echo "alias update-passwordmanager='cd ~/SourceVersionControls/github.com/virtualdemon/passwordmanager && git pull -f && cd'"
+    echo "alias passwordmanager='cd ~/.passwordmanager && activenv && ./passwordmanager.py'"
+    echo "alias update-passwordmanager='cd ~/.passwordmanager && git pull -f && cd'"
     exit
 fi
 # activate virtual environment for installing the needed libraries
